Modifications post code review, some tidy up of unused properties and renaming of files to follow conventions
