+++ projects/media-viewer/src/lib/shared/action-events.ts
@@ -15,7 +15,7 @@ import {
 
 /**
  * @deprecated Please DO NOT ADD to this file
- * Instead please add new Operations to toolbar.events.service.ts
+ * Instead please add new Operations to toolbar-events.service.ts
  */
 export class ActionEvents {
   public readonly rotate = new Subject<RotateOperation>();
@@ -28,5 +28,4 @@ export class ActionEvents {
   public readonly download = new Subject<DownloadOperation>();
   public readonly setCurrentPage = new Subject<SetCurrentPageOperation>();
   public readonly changePageByDelta = new Subject<ChangePageByDeltaOperation>();
-  public readonly highlightMode = new BehaviorSubject<boolean>(false);
 }
