+++ projects/media-viewer/src/lib/shared/toolbar-button-toggles.ts
@@ -15,7 +15,6 @@ export class ToolbarButtonToggles {
   public readonly subToolbarHidden = new BehaviorSubject(true);
   public readonly sidebarOpen = new BehaviorSubject(false);
   public readonly searchBarHidden = new BehaviorSubject(true);
-  public readonly drawMode = new BehaviorSubject(false);
 
   public showPrintBtn = false;
   public showDownloadBtn = false;
