+++ projects/media-viewer/src/lib/shared/viewer-events.service.ts
