+++ projects/media-viewer/src/lib/toolbar/main-toolbar/main-toolbar.component.ts
@@ -31,7 +31,6 @@ export class MainToolbarComponent {
   @Input() rotateEvent: Subject<RotateOperation>;
   @Input() changePageByDeltaEvent: Subject<ChangePageByDeltaOperation>;
   @Input() setCurrentPageEvent: Subject<SetCurrentPageOperation>;
-  @Input() highlightMode: BehaviorSubject<boolean>;
   @Input() downloadEvent: Subject<DownloadOperation>;
   @Input() printEvent: Subject<PrintOperation>;
 }
